+++ printing/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.4
 * Fix iOS build with Swift
 * Add installation instructions in the Readme
+* Follow Flutter debug painting settings
 
 # 1.3.3
 * Fix dart lint warnings